Add data class and parse data vars
